requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
inflect==7.5.0
alembic==1.15.2
pydantic==2.11.4
passlib==1.7.4
pydantic[email]
bcrypt==4.3.0
python-slugify==8.0.4